requirements.txt
aiocron==2.1
aiofiles==24.1.0
anyio==4.9.0
APScheduler==3.11.0
asyncio==3.4.3
asyncpg==0.30.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cronsim==2.6
dotenv==0.9.9
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
numpy==2.2.4
oauthlib==3.2.2
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pygsheets==2.0.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
